Clean up piping reviews
Problem:

The filesc chan was never closed for a tenet. Because the tenet map was being channelled in, there was no way to tell if the filesc for that tenet could safely be closed - as a matching cfgHash for that tenet could still be met.

The pendingReview added nothing but to hide the above fact. The fact that tenets were closing at all was a side effect. They relied on the no files timeout in lingo/tenet/tenet.go to close the filesc chan of every tenet. This explains why the binary stress test always took 10 seconds for 13k or 20k issues. lingo didn't blow up because all reviews where already complete and it was okay to close the chan.

Docker does blow up because it takes longer than 10sec to finish the review. If you reviewed with docker, it just cuts you off after 10sec.

Solution:

- get rid of pendingReview
- introduce a tomb
- filesc is stopped one of two ways:
  - 1. all reviews are complete
  - 2. no file has been fed to the tenet in 5seconds
- add a wait group to ensure all tenets are cleaned up

The total number of issues found was different every time, this is now fixed.

See inline comments for more details. An important concept is that the take down of a tenet should start with the closing of filesc - this was completely missing from the code.

It is now stable on large docker reviews. Two outstanding bits of work left:

1. close all tenets on sigterm
2. throttle the background review. If the user has one issue up, their cpu + mem should not be maxing out.