Service Sub Commands
This branch introduces sub commands to `lingo review` to send the result of a
review to a service. Configuration for each service is written in
$LINGO_HOME/services.yaml

The branch also introduces the `lingo copr` command which stands for "checkout
pull request". run `lingo copr --help` to see how it works.

A third command `lingo update` was added. This is to be run after the lingo
client tool has been upgraded. The command updates any required resources in
$LINGO_HOME.

As part of the refactoring, commands/common.go was moved into its
own package. The PR also contains several drive-by bug fixes and code clean-ups.